Add ColorValidate extension
